Fixed gcc warnings, general (un)signed usage cleanup
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@27508 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/common/classes/MetaName.cpp
@@ -103,7 +103,7 @@ namespace Firebird {
 		va_list params;
 		va_start(params, format);
 		int l = VSNPRINTF(data, MAX_SQL_IDENTIFIER_LEN, format, params);
-		if (l < 0 || l > MAX_SQL_IDENTIFIER_LEN)
+		if (l < 0 || size_t(l) > MAX_SQL_IDENTIFIER_LEN)
 		{
 			l = MAX_SQL_IDENTIFIER_LEN;
 		}
